mqtt-pubsub/src/requirements.txt
paho-mqtt==1.5.0
influxdb==5.3.0
